Add Image Source Swift Package
