--- GetMoovin/PhotoUpload.swift
@@ -1,21 +1,0 @@
-//
-// This source file is part of the CS342 2023 GetMoovin Team Application project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-import SwiftUI
-
-struct PhotoUpload: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct PhotoUpload_Previews: PreviewProvider {
-    static var previews: some View {
-        PhotoUpload()
-    }
-}
